[GWT]: Scripting performance.

+++ platform/web-client/src/main/java/platform/gwt/form/client/dispatch/QueuedAction.java
@@ -41,7 +41,7 @@ public class QueuedAction<R extends Result> {
         }
     }
 
-    public long getRequestIndex() {
+    public int getRequestIndex() {
         return action instanceof FormRequestIndexCountingAction ? ((FormRequestIndexCountingAction) action).requestIndex : -1;
     }
 }
