+++ platform/web-client/src/main/java/platform/gwt/form/shared/actions/form/FormRequestIndexCountingAction.java
@@ -3,7 +3,7 @@ package platform.gwt.form.shared.actions.form;
 import net.customware.gwt.dispatch.shared.Result;
 
 public class FormRequestIndexCountingAction<R extends Result> extends FormBoundAction<R> {
-    public long requestIndex;
+    public int requestIndex;
 
     public FormRequestIndexCountingAction() {
     }
